feature(frontend): componentizing frontend and added menu
componentizing frontend and added menu
